requirements.txt
torch>=2.0
torchvision>=0.15
scikit-learn>=1.3
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.65
Pillow>=9.0
